Add ability to list files for which warning 4 is disabled
Since we'd like to enable the 'fragile pattern matches' warning, but
it's hard to get rid of it in some files (mostly pattern-matching on
`Unix.error`), we need a way to disable to corresponding warning for
these specific files.

This change adds the feature to `myocamlbuild.ml`, not whitelisting any
file for now.